Refactoring math_tools headerfile
