+++ src/game_logic/enemies/boss_episode_4.cpp
@@ -16,7 +16,7 @@
 
 #include "boss_episode_4.hpp"
 
-#include "base/math_tools.hpp"
+#include "base/math_utils.hpp"
 #include "engine/base_components.hpp"
 #include "engine/life_time_components.hpp"
 #include "engine/physical_components.hpp"
