+++ src/game_logic/enemies/laser_turret.cpp
@@ -16,7 +16,7 @@
 
 #include "laser_turret.hpp"
 
-#include "base/math_tools.hpp"
+#include "base/math_utils.hpp"
 #include "common/game_service_provider.hpp"
 #include "data/player_model.hpp"
 #include "engine/random_number_generator.hpp"
