+++ src/game_logic/enemies/rocket_turret.cpp
@@ -16,7 +16,7 @@
 
 #include "rocket_turret.hpp"
 
-#include "base/math_tools.hpp"
+#include "base/math_utils.hpp"
 #include "common/game_service_provider.hpp"
 #include "engine/base_components.hpp"
 #include "engine/sprite_tools.hpp"
