+++ src/game_logic/player.cpp
@@ -17,7 +17,7 @@
 #include "player.hpp"
 
 #include "base/match.hpp"
-#include "base/math_tools.hpp"
+#include "base/math_utils.hpp"
 #include "common/game_service_provider.hpp"
 #include "data/game_options.hpp"
 #include "data/game_traits.hpp"
