+++ src/ui/duke_script_runner.cpp
@@ -18,7 +18,7 @@
 
 #include "base/container_utils.hpp"
 #include "base/match.hpp"
-#include "base/math_tools.hpp"
+#include "base/math_utils.hpp"
 #include "common/game_service_provider.hpp"
 #include "data/game_traits.hpp"
 #include "data/unit_conversions.hpp"
